+++ src/main/java/pl/plajer/buildbattle/handlers/language/LanguageMigrator.java
@@ -41,7 +41,7 @@ import pl.plajerlair.commonsbox.minecraft.migrator.MigratorUtils;
 public class LanguageMigrator {
 
   public static final int LANGUAGE_FILE_VERSION = 15;
-  public static final int CONFIG_FILE_VERSION = 9;
+  public static final int CONFIG_FILE_VERSION = 10;
   private List<String> migratable = Arrays.asList("bungee", "config", "language", "mysql");
   private Main plugin;
 
@@ -185,6 +185,14 @@ public class LanguageMigrator {
               "    Full-Game: \"&4&lFULL\"\r\n" +
               "    Ending: \"&lEnding\"\r\n" +
               "    Restarting: \"&c&lRestarting\"\r\n");
+          break;
+        case 8:
+          //Updating to new config file for gtb
+          break;
+        case 9:
+          MigratorUtils.addNewLines(file, "\r\n" +
+                  "# Should we enable short commands such as /start and /leave\r\n" +
+                  "Enable-Short-Commands: false\r\n");
           break;
       }
     }