+++ CHANGELOG.md
@@ -12,6 +12,7 @@ That's all, matcher will stop when detects next line started with `###` match
 
 ### 4.4.0 Release
 * GuessTheBuild release
+* Added short commands such as start and leave
 * Tried to fix Server crash on 1.15 
 * Reworked whole Rewards
 * New file rewards.yml